+++ src/controllers/firewalls/openvpn/archive/config/dtos/openvpn-archive-config-update.dto.ts
@@ -1,0 +1,11 @@
+import { IsNumber, IsPositive } from "class-validator";
+
+export class OpenVPNArchiveControllerUpdateDto {
+    @IsNumber()
+    @IsPositive()
+    archive_days: number;
+    
+    @IsNumber()
+    @IsPositive()
+    retention_days: number;
+}