+++ src/middleware/InputValidation.ts
@@ -43,10 +43,11 @@ export class InputValidation extends Middleware {
         const item1 = req.url.split('/')[1];
         const item1_valid_list = ['user', 'customer', 'fwcloud', 'firewall', 'cluster', 'policy', 'interface', 'ipobj', 'tree', 'vpn'];
 
-        const item1_new_route_system = ['backups', 'version', 'fwclouds', 'updates', 'iptables-save', 'ping', 'profile'];
+        const item1_new_route_system = ['backups', 'version', 'fwclouds', 'updates', 'iptables-save', 'ping', 'profile', 'openvpnarchives'];
 
         // Verify that item1 is in the valid list.
         if (!item1_valid_list.includes(item1) && !item1_new_route_system.includes(item1.replace(/\?.*/, ''))) {
+            logger().error('Unknown route: ' + item1)
             logger().error('Error during input validation check: ' + JSON.stringify(fwcError.BAD_API_CALL));
             res.status(404).json(fwcError.BAD_API_CALL);
             return;